coro: sync latest xgo-dev main

+++ cl/_testrt/strlen/in.go
@@ -1,4 +1,3 @@
-// LITTEST
 package main
 
 import "C"
@@ -8,17 +7,11 @@ import _ "unsafe"
 func printf(format *int8, __llgo_va_list ...any)
 
 //go:linkname strlen C.strlen
-func strlen(str *int8) C.int
+func strlen(str *int8) uintptr
 
 var format = [...]int8{'H', 'e', 'l', 'l', 'o', ' ', '%', 'd', '\n', 0}
 
-// CHECK-LABEL: define void @main.main(){{.*}} {
-// CHECK-NEXT: _llgo_0:
-// CHECK-NEXT:   %0 = call i32 @strlen(ptr @main.format)
-// CHECK-NEXT:   call void (ptr, ...) @printf(ptr @main.format, i32 %0)
-// CHECK-NEXT:   ret void
-// CHECK-NEXT: }
 func main() {
 	sfmt := &format[0]
-	printf(sfmt, strlen(sfmt))
+	printf(sfmt, C.int(strlen(sfmt)))
 }
